+++ src/main/java/com/semiuniv/semiu/repository/ClassroomRepository.java
@@ -1,23 +1,12 @@
 package com.semiuniv.semiu.repository;
 
 import com.semiuniv.semiu.entity.Classroom;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import jakarta.transaction.Transactional;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-@Repository
-@Transactional
-public class ClassroomRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    public List<Integer> findAllIds() {
-        return em.createQuery("SELECT c.id FROM Classroom c", Integer.class)
-                .getResultList();
-    }
-
+public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
+    @Query(value = "SELECT classroom_id FROM Classroom ", nativeQuery = true)
+    List<Integer> findAllIds();
 }
