+++ src/main/java/com/semiuniv/semiu/repository/UserRepository.java
@@ -6,4 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<Users, String> {
+
 }
