+++ src/main/resources/templates/subjects/insertSubject.html
@@ -5,7 +5,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Subject Insert</title>
-    <script type="text/javascript" th:src="@{/js/subjectCheck.js}"></script>
+    <script type="text/javascript" th:src="@{/js/subjectCheck2.js}"></script>
 </head>
 <body>
     <div class="content" layout:fragment="content">
@@ -27,11 +27,6 @@
                 <span id="selectedProfessorName"></span>
             </div>
 
-
-<!--            <div class="border border-top-0">-->
-<!--                <label for="professorName" class="col-sm-4 col-form-label border border-top-0 border-bottom-0 border-left-0" style="text-align: center;" >교수</label>-->
-<!--                <input type="text" id="professorName" th:field="*{professor.name}">-->
-<!--            </div>-->
             <div class="border border-top-0">
                 <label class="col-sm-4 col-form-label border border-top-0 border-bottom-0 border-left-0"
                        style="text-align: center;" >강의실</label>
