+++ src/main/resources/templates/subjects/updateSubject.html
@@ -5,7 +5,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Subject Update</title>
-    <script type="text/javascript" th:src="@{/js/subjectCheck.js}"></script>
+    <script type="text/javascript" th:src="@{/js/subjectCheck2.js}"></script>
 </head>
 <body>
     <div class="content" layout:fragment="content">
@@ -17,14 +17,16 @@
                 <label for="name" class="col-sm-4 col-form-label border border-top-0 border-bottom-0 border-left-0" style="text-align: center;" >강의명</label>
                 <input type="text" id="name" th:field="*{name}">
             </div>
-            <div class="border border-top-0">
-                <label for="professorName" class="col-sm-4 col-form-label border border-top-0 border-bottom-0 border-left-0" style="text-align: center;" >교수</label>
-                <input type="text" id="professorName" th:field="*{professor.name}">
+            <div class="border border-top-0" >
+                <label class="col-sm-4 col-form-label border border-top-0 border-bottom-0 border-left-0"
+                       style="text-align: center;" >교수</label>
+                <select th:field="*{professor.id}"  id="professor" onchange="professorCheck()">
+                    <option value="">선택</option>
+                    <option th:each="professor : ${professors}" th:value="${professor.id}"
+                            th:text="${professor.id}" th:data-name="${professor.name}"></option>
+                </select>
+                <span id="selectedProfessorName"></span>
             </div>
-<!--            <div class="border border-top-0">-->
-<!--                <label for="classroomId" class="col-sm-4 col-form-label border border-top-0 border-bottom-0 border-left-0" style="text-align: center;" >강의실</label>-->
-<!--                <input type="text" id="classroomId" th:field="*{classroom.id}">-->
-<!--            </div>-->
             <div class="border border-top-0">
                 <label class="col-sm-4 col-form-label border border-top-0 border-bottom-0 border-left-0"
                        style="text-align: center;" >강의실</label>
